Rename audio.h to fa_audio.h
