+++ apps/freeablo/faaudio/audiomanager.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <audio/audio.h>
+#include <audio/fa_audio.h>
 #include <list>
 #include <map>
 #include <stdint.h>
